+++ ChangeLog.md
@@ -18,6 +18,7 @@
 - Automatically sort alpha on cards (actors, locations), enable disable in settings
 - Create and link a location by dropping a scene onto the location tab of a region, scene links to location. Checks for duplicate (same loc name, same scene) and won't create.
 - Fixed names not updating on open linked journals until reopen
+- Link a journal to an info tab of a sheet or group sheet
 
 ## 0.7.5
 - Moved to GitHub